requirements.txt
anyio==4.3.0
    # via httpx
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
h11==0.14.0
    # via httpcore
httpcore==1.0.3
    # via httpx
httpx==0.26.0
    # via python-telegram-bot
idna==3.6
    # via
    #   anyio
    #   httpx
python-dotenv==1.0.0
python-telegram-bot==20.8
apscheduler>=3.10.0
aiosqlite>=0.19.0
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
